Module Bits.Expert

val unsafe_underlying_repr : t -> Base.bytes

Access the underlying data representation. Note that this is unstable, and may change over time.

val offset_for_data : Base.int

Offset to access actual data within in the underlying repr.