Module Hardcaml.Bits_list

Combinational logic described as lists of Bits ie 0;1;1;1;0.

Width implicitly defined as the length of the list.

module type T = sig ... end

module type Comb = sig ... end

module Make (T : T) : Comb with type t = T.t Base.list

From T construct the combinational API

module Int_comb : Comb with type t = Base.int Base.list

Uses 0,1 as the bit type.

module Bool_comb : Comb with type t = Base.bool Base.list

Uses false,true as the bit type.

module X : sig ... end

Logic with undefined values.

module X_comb : Comb with type t = X.t Base.list

Uses F, T, X as the bit type.