Module Bits_list.X

Logic with undefined values.

type t =

| F

| T

| X

include Ppx_compare_lib.Comparable.S with type t := t

val compare : t Base__Ppx_compare_lib.compare

include Ppx_compare_lib.Equal.S with type t := t

val equal : t Base__Ppx_compare_lib.equal

val sexp_of_t : t -> Sexplib0.Sexp.t

include T with type t := t

include Ppx_compare_lib.Equal.S with type t := t

val equal : t Base__Ppx_compare_lib.equal

include Sexplib0.Sexpable.S with type t := t

val t_of_sexp : Sexplib0__.Sexp.t -> t

val sexp_of_t : t -> Sexplib0__.Sexp.t

val constant_only : Base.bool

Only constants are valid (ie the type t does not represent a notion of inputs). This allows more compact sexp printing.

val optimise_muxs : Base.bool

Mux's should be optimised. The select values will be converted to an int and used access the data list. Only valid in conjunction with constant_only.

val vdd : t

val gnd : t

val (&:) : t -> t -> t

val (|:) : t -> t -> t

val (^:) : t -> t -> t

val (~:) : t -> t

val to_char : t -> Base.char

val of_char : Base.char -> t