Module Hardcaml.Caller_id

Embed the callstack in the signal type when it is created.

When an exception occurs, the callstack is printed as part of the sexp of the signal. This is especially useful for dangling wires as it shows the construction site of the wire rather than the place that the dangling wire was detected.

By default it is not enabled.

module Mode : sig ... end

type t

val sexp_of_t : t -> Sexplib0.Sexp.t

val set_mode : Mode.t -> Base.unit

val get : ?skip:Base.string Base.list -> Base.unit -> t Base.option