Module Circuit.Port_checks

Check if the ports specified in the interface match those defined in the circuit.

type t =

| Relaxed	(*
No checks
*)

| Port_sets	(*
Input and output port sets agree
*)

| Port_sets_and_widths	(*
Input and output port sets agree, and their widths are the same.
*)