Module Hardcaml.Combinational_ops_database

A database which holds a collecton of custom combinational operations for use with Cyclesim based simulators.

type t

val sexp_of_t : t -> Sexplib0.Sexp.t

val create : Base.unit -> t

Create an empty database.

val insert : t -> Combinational_op.t -> Base.unit

Insert combinational op into the database.

val find : t -> name:Base.string -> Combinational_op.t Base.option

Find a combinational op in the database, given its name.

val fold : t -> init:'a -> f:( 'a -> Combinational_op.t -> 'a ) -> 'a

val iter : t -> f:( Combinational_op.t -> Base.unit ) -> Base.unit

val concat : t Base.list -> t

concat ts creates a new database with all elements added from the list of databases ts.