Module Interface.Ast

module Ast : sig ... end

The PPX can optionally generate an ast field containing an Ast.t. This represents the structure of the interface, including how it is constructed from fields, arrays, lists and sub-modules.

module Field : sig ... end

module Type : sig ... end

module Sequence : sig ... end

type t = Ast.t

val sexp_of_t : t -> Sexplib0.Sexp.t