Module Hardcaml.Property_manager

type t

val sexp_of_t : t -> Sexplib0.Sexp.t

val create : Base.unit -> t

val add_ltl : t -> Base.string -> Property.LTL.path -> Base.unit

Add a named LTL property

val finalize : t -> Property.LTL.path Base.Map.M(Base.String).t

Disables adding new properties, and returns a map of the LTL properties added. Calling it again will return the map again

val atomic_propositions : t -> Signal.t Base.list

Finalizes the Property_manager if not already finalized, and return a list of all Signal.t used as atomic propositions. This list will not contain duplicates