Module Hardcaml.Rtl_attribute

RTL attribute specification. Only relevant to downstream tooling.

type t

Specification of attributes which may be attached to various objects within a RTL design. Such attributes are used to provide implementation hints to down stream CAD tools and do not affect any functionality within Hardcaml.

val sexp_of_t : t -> Sexplib0.Sexp.t

module Value : sig ... end

Attribute value types.

val create : ?value:Value.t -> string -> t

Create a new attribute.

val name : t -> string

Returns the attribute name

val value : t -> Value.t option

Returns the attribute value, if any.

module Vivado : sig ... end

A collection of common Xilinx Vivado attributes.