Module Hardcaml.Transform

Transform circuits to a different representation.

type 'a transform_fn' = ( Signal.Uid.t -> 'a ) -> Signal.t -> 'a

type transform_fn = Signal.t transform_fn'

module type TransformFn' = sig ... end

module type TransformFn = sig ... end

module type MakePureCombTransform_arg = sig ... end

module MakePureCombTransform (B : MakePureCombTransform_arg) : TransformFn' with type t = B.t

functor to build the function to map a signal to a new combinatorial signal representation

module MakeCombTransform (B : Comb.Primitives with type t = Signal.t) : TransformFn

functor to build the function to map a signal to a new combinatorial signal representation

module CopyTransform : TransformFn

simple copying transform

val rewrite_signals : transform_fn -> Signal.t Base.list -> Signal.t Base.list

rewrites the list of signals based on the given function